+++ src/multi_bot_agentic/tools/__init__.py
@@ -29,6 +29,7 @@ from multi_bot_agentic.tools.html_markdown import HtmlMarkdownTool
 from multi_bot_agentic.tools.html_strip import HtmlStripTool
 from multi_bot_agentic.tools.html_table import HtmlTableTool
 from multi_bot_agentic.tools.html_table_csv import HtmlTableCsvTool
+from multi_bot_agentic.tools.json_flatten import JsonFlattenTool
 from multi_bot_agentic.tools.json_format import JsonFormatTool
 from multi_bot_agentic.tools.json_merge_patch import JsonMergePatchTool
 from multi_bot_agentic.tools.json_path import JsonPathTool
@@ -96,6 +97,7 @@ __all__ = [
     "HtmlStripTool",
     "HtmlTableCsvTool",
     "HtmlTableTool",
+    "JsonFlattenTool",
     "JsonFormatTool",
     "JsonMergePatchTool",
     "JsonPathTool",